refactoring navbar in layout
